requirements.txt
absl-py==0.13.0
aiohttp==3.8.0
aiosignal==1.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.0
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
bidict==0.22.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.5
clang==5.0
click==8.0.1
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docstring-parser==0.14.1
editdistance==0.6.0
einops==0.3.2
entrypoints==0.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2021.10.1
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
imageio==2.19.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jsonargparse==4.9.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.18.0
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytorch-lightning==1.4.9
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.6.0
torchmetrics==0.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
yarl==1.7.0
zipp==3.5.0


